Master

+++ service/src/main/java/greencity/security/service/PasswordRecoveryServiceImpl.java
@@ -1,29 +1,33 @@
 package greencity.security.service;
 
 import greencity.constant.ErrorMessage;
+import greencity.entity.OwnSecurity;
 import greencity.entity.RestorePasswordEmail;
 import greencity.entity.User;
 import greencity.enums.UserStatus;
-import greencity.exception.exceptions.BadVerifyEmailTokenException;
+import greencity.exception.exceptions.BadRequestException;
 import greencity.exception.exceptions.NotFoundException;
 import greencity.exception.exceptions.UserActivationEmailTokenExpiredException;
 import greencity.exception.exceptions.WrongEmailException;
-import greencity.message.PasswordRecoveryMessage;
 import greencity.repository.UserRepo;
+import greencity.security.dto.ownsecurity.OwnRestoreDto;
 import greencity.security.events.UpdatePasswordEvent;
 import greencity.security.jwt.JwtTool;
+import greencity.security.repository.OwnSecurityRepo;
 import greencity.security.repository.RestorePasswordEmailRepo;
+import greencity.service.EmailService;
+import java.time.LocalDateTime;
+import java.util.Optional;
+
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDateTime;
-
-import static greencity.constant.RabbitConstants.PASSWORD_RECOVERY_ROUTING_KEY;
+import javax.persistence.NoResultException;
 
 /**
  * Service for password recovery functionality. It manages recovery tokens
@@ -34,38 +38,41 @@ import static greencity.constant.RabbitConstants.PASSWORD_RECOVERY_ROUTING_KEY;
 @Service
 @Slf4j
 public class PasswordRecoveryServiceImpl implements PasswordRecoveryService {
+    private final OwnSecurityRepo ownSecurityRepo;
+    private final PasswordEncoder passwordEncoder;
     private final UserRepo userRepo;
     private final RestorePasswordEmailRepo restorePasswordEmailRepo;
     private final ApplicationEventPublisher applicationEventPublisher;
-    private final RabbitTemplate rabbitTemplate;
     private final JwtTool jwtTool;
+    private final EmailService emailService;
     @Value("${verifyEmailTimeHour}")
     private Integer tokenExpirationTimeInHours;
-    @Value("${messaging.rabbit.email.topic}")
-    private String sendEmailTopic;
 
     /**
      * Constructor with all essentials beans for password recovery functionality.
      *
+     * @param ownSecurityRepo           - security repository.
+     * @param passwordEncoder           - encodes password.
      * @param restorePasswordEmailRepo  {@link RestorePasswordEmailRepo} - Used for
      *                                  storing recovery tokens
      * @param applicationEventPublisher {@link ApplicationEventPublisher} - Used for
      *                                  publishing events, such as email sending or
      *                                  password update
-     * @param rabbitTemplate            template for sending RabbitMQ messages
      * @param jwtTool                   {@link JwtTool} - Used for recovery token
-     *                                  generation
      */
     public PasswordRecoveryServiceImpl(
+        OwnSecurityRepo ownSecurityRepo, PasswordEncoder passwordEncoder,
         RestorePasswordEmailRepo restorePasswordEmailRepo,
         UserRepo userRepo,
         ApplicationEventPublisher applicationEventPublisher,
-        RabbitTemplate rabbitTemplate,
+        EmailService emailService,
         JwtTool jwtTool) {
+        this.ownSecurityRepo = ownSecurityRepo;
+        this.passwordEncoder = passwordEncoder;
         this.restorePasswordEmailRepo = restorePasswordEmailRepo;
         this.userRepo = userRepo;
         this.applicationEventPublisher = applicationEventPublisher;
-        this.rabbitTemplate = rabbitTemplate;
+        this.emailService = emailService;
         this.jwtTool = jwtTool;
     }
 
@@ -82,7 +89,7 @@ public class PasswordRecoveryServiceImpl implements PasswordRecoveryService {
         if (restorePasswordEmail != null) {
             throw new WrongEmailException(ErrorMessage.PASSWORD_RESTORE_LINK_ALREADY_SENT + email);
         }
-        savePasswordRestorationTokenForUser(user, jwtTool.generateTokenKey(), language);
+        savePasswordRestorationTokenForUser(user, jwtTool.generateTokenKeyWithCodedDate(), language);
     }
 
     /**
@@ -90,20 +97,24 @@ public class PasswordRecoveryServiceImpl implements PasswordRecoveryService {
      */
     @Transactional
     @Override
-    public void updatePasswordUsingToken(String token, String newPassword) {
+    public void updatePasswordUsingToken(OwnRestoreDto form) {
         RestorePasswordEmail restorePasswordEmail = restorePasswordEmailRepo
-            .findByToken(token)
-            .orElseThrow(() -> new BadVerifyEmailTokenException(ErrorMessage.NO_ANY_EMAIL_TO_VERIFY_BY_THIS_TOKEN));
+            .findByToken(form.getToken())
+            .orElseThrow(() -> new NotFoundException(ErrorMessage.TOKEN_FOR_RESTORE_IS_INVALID));
+        if (!form.getPassword().equals(form.getConfirmPassword())) {
+            throw new BadRequestException(ErrorMessage.PASSWORDS_DO_NOT_MATCH);
+        }
         UserStatus userStatus = restorePasswordEmail.getUser().getUserStatus();
         if (isNotExpired(restorePasswordEmail.getExpiryDate())) {
+            updatePassword(form.getPassword(), restorePasswordEmail.getUser().getId());
             applicationEventPublisher.publishEvent(
-                new UpdatePasswordEvent(this, newPassword, restorePasswordEmail.getUser().getId()));
+                new UpdatePasswordEvent(this, form.getPassword(), restorePasswordEmail.getUser().getId()));
             restorePasswordEmailRepo.delete(restorePasswordEmail);
             log.info("User with email " + restorePasswordEmail.getUser().getEmail()
-                + " has successfully restored the password using token " + token);
+                + " has successfully restored the password using token " + form.getToken());
         } else {
             log.info("Password restoration token of user with email " + restorePasswordEmail.getUser().getEmail()
-                + " has been expired. Token: " + token);
+                + " has been expired. Token: " + form.getToken());
             throw new UserActivationEmailTokenExpiredException(ErrorMessage.EMAIL_TOKEN_EXPIRED);
         }
         if (userStatus == UserStatus.CREATED) {
@@ -126,15 +137,12 @@ public class PasswordRecoveryServiceImpl implements PasswordRecoveryService {
                 .expiryDate(calculateExpirationDate(tokenExpirationTimeInHours))
                 .build();
         restorePasswordEmailRepo.save(restorePasswordEmail);
-        rabbitTemplate.convertAndSend(
-            sendEmailTopic,
-            PASSWORD_RECOVERY_ROUTING_KEY,
-            new PasswordRecoveryMessage(
-                user.getId(),
-                user.getFirstName(),
-                user.getEmail(),
-                token,
-                language));
+        emailService.sendRestoreEmail(
+            user.getId(),
+            user.getFirstName(),
+            user.getEmail(),
+            token,
+            language);
     }
 
     /**
@@ -170,5 +178,22 @@ public class PasswordRecoveryServiceImpl implements PasswordRecoveryService {
     void deleteAllExpiredPasswordResetTokens() {
         int rows = restorePasswordEmailRepo.deleteAllExpiredPasswordResetTokens();
         log.info(rows + " password reset tokens were deleted.");
+    }
+
+    private void updatePassword(String pass, Long id) {
+        String password = passwordEncoder.encode(pass);
+        Optional<OwnSecurity> ownSecurity = ownSecurityRepo.findByUserId(id);
+
+        ownSecurity.ifPresentOrElse(s -> {
+            s.setPassword(password);
+            ownSecurityRepo.save(s);
+        }, () -> ownSecurityRepo.save(createOwnSecurity(id, password)));
+    }
+
+    private OwnSecurity createOwnSecurity(Long id, String password) {
+        return OwnSecurity.builder()
+            .password(password)
+            .user(userRepo.findById(id).orElseThrow(NoResultException::new))
+            .build();
     }
 }
